refactor deriving all accent color variants into a function
