+++ frontend/src/metabase/lib/colors/derive-theme.ts
@@ -5,11 +5,7 @@ import type { ColorSettings } from "metabase-types/api";
 
 import type { ResolvedColorScheme } from "../color-scheme";
 
-import {
-  deriveChartShadeColor,
-  deriveChartTintColor,
-  mapChartColorsToAccents,
-} from "./accents";
+import { deriveAllAccentColors, mapChartColorsToAccents } from "./accents";
 import { PROTECTED_COLORS } from "./constants/protected-colors";
 import { getThemeFromColorScheme } from "./theme-from-color-scheme";
 import type { MetabaseColorKey, MetabaseDerivedThemeV2 } from "./types";
@@ -42,54 +38,12 @@ export function deriveFullMetabaseTheme({
     ...PROTECTED_COLORS,
   );
 
-  const whitelabelAccentColors = Object.fromEntries(
-    Object.entries(whitelabelColors ?? {}).filter(([key]) =>
-      key.startsWith("accent"),
-    ),
-  );
-  const allWhitelabelAccentColors = Object.entries(
-    whitelabelAccentColors,
-  ).reduce((acc, [key, value]) => {
-    const [accent, variant] = key.split("-");
-    let colorsToMaybeSet;
-    if (variant === "light") {
-      const base = deriveChartShadeColor(value);
-      colorsToMaybeSet = {
-        [accent]: base,
-        [`${accent}-dark`]: deriveChartShadeColor(base),
-      };
-    } else if (variant === "dark") {
-      const base = deriveChartTintColor(value);
-      colorsToMaybeSet = {
-        [accent]: base,
-        [`${accent}-light`]: deriveChartTintColor(base),
-      };
-    } else {
-      colorsToMaybeSet = {
-        [`${accent}-light`]: deriveChartTintColor(value),
-        [`${accent}-dark`]: deriveChartShadeColor(value),
-      };
-    }
-    return {
-      // acc takes priority because it's initialized with user provided whitelabelAccentColors
-      ...colorsToMaybeSet,
-      ...acc,
-    };
-  }, whitelabelAccentColors);
-
-  const whitelabelOtherColors = Object.fromEntries(
-    Object.entries(whitelabelColors ?? {}).filter(
-      ([color]) => !color.startsWith("accent"),
-    ),
-  );
-
   return {
     version: 2,
     colors: {
       ...baseTheme.colors,
       ...mapChartColorsToAccents(baseTheme.chartColors),
-      ...allWhitelabelAccentColors,
-      ...whitelabelOtherColors,
+      ...deriveAllAccentColors(whitelabelColors ?? {}),
       ...filteredEmbeddingColors,
       ...mapChartColorsToAccents(embeddingThemeOverride?.chartColors ?? []),
     } as Record<MetabaseColorKey, string>,
